Using "--format" with  log() now throws exception. [#43]
Issue 43: Updated Git::Wrapper::log to throw a Git::Wrapper::Exception
if "--format" is detected as a passed-in option. A unit test was added
to make sure the exception was thrown in the right case. The POD was
updated to underscore the fact that the log method shall die if
"--format" is detected and points reader to the documentation on the RUN
method, which now includes an example of calling 'log' directly with the
'--format' option.