Add failing test
I think to be consistent with the rest of the interface, the
Git::Wrapper::RawModification object for a change with a score should
return the current file name.

I'd be open to other suggestions as to how to handle this, but returning
both old and new names separated by whitespace seems like not the right
thing.